refactored services and users
